refactor(frontend): extract 3 small components (4/N App.jsx split)
Moves three pure presentational components out of App.jsx:
- components/BrandLogo.jsx       — SVG logo (the magnifying glass)
- components/LogoImg.jsx         — Company logo with favicon→Clearbit→
                                    initial-tile fallback chain. Bundled
                                    with its CO_DOMAINS table + guessDomain
                                    helper since they're only used here.
- components/ScrapeProgressBar.jsx — Animated progress bar for the
                                      Monitor tab + admin op cards.

No behavior change. Build clean.

App.jsx: -140 LOC.
